move cuda test to a separate file
